// src/app/src_services_commentes/src/app/interfaces/post.ts
import { User } from 'src/app/interfaces/user';
export interface Post {
    _id: string;
    title: string;
    sub_title: string;
    categorie: string;
    partner: string;
    doc: string;
    categorie_id: any;
    images: any;
    video: any;
    pdf: any;
    chambre: number;
    seen: number;
    salon: number;
    toilette: number;
    couloir: number;
    terrasse: number;
    cuisine: number;
    sale_bain: number;
    sale_manger: number;
    sole: string;
    parcking: boolean;
    meuble: boolean;
    garage: boolean;
    nb_day_free: number;
    mode_payement: boolean;
    climatiseur: boolean;
    lotissement: boolean;
    discut: boolean;
    terassement: boolean;
    mettreCaree: number;
    prix: number;
    avance: any;
    etat_bien: string;
    periode: string;
    caution: number;
    mont_invest: number;
    benefice_prevue: number;
    terms: any;
    temps: number;
    cour: boolean;
    etage: boolean;
    gardien: boolean;
    active: boolean;
    region: string;
    eaux: string;
    contact: any[];
    nivauxEtage: string;
    commune: string;
    quartier: string;
    secteur: string;
    description: string;
    detail_construction: string;
    condition: string;
    fournisseur: string;
    delete: number;
    postedBy: number;
    status_devi: number;
    views: number;
    ing: boolean;
    starts: number;
    postId: number;
    status: number;
    type: number;
    devis: any;
    date: Date;
    createdAt: Date;
    user_id: User;
  }